Change type of conidx from u32 to u8

+++ bitbox02-bt/src/ble.rs
@@ -20,7 +20,7 @@ impl BleDriver for Da14531Ble {
         app_easy_gap_advertise_stop();
     }
 
-    fn disconnect(connection_handle: u32) {
-        app_easy_gap_disconnect(connection_handle as u8);
+    fn disconnect(connection_handle: u8) {
+        app_easy_gap_disconnect(connection_handle);
     }
 }
